requirements.txt
telethon
groq
flask
fpdf2
python-dotenv
